[#133] confirm delete function

+++ src/app/nation/nation.component.ts
@@ -64,10 +64,13 @@ export class NationComponent implements OnInit {
   }
 
   deleteNation(nationId, nationRev, index){
+    const nationDelete = confirm('Are you sure you want to delete?')
+    if (nationDelete){
       this.couchService.delete('nations/' + nationId + '?rev=' + nationRev)
       .then((data) => {
         this.nation.splice(index,1);
       }, (error) => this.message = 'There was a problem deleting this meetup');
+    }
   }
 
   onSubmit(nation) {
